fpx to f64

+++ packages/frameworks/rmf-static/src/audio/utils.rs
@@ -1,13 +1,13 @@
 #[inline]
-pub fn calculate_frame_samples(fps: f32, sample_rate: u32, position: isize) -> isize {
+pub fn calculate_frame_samples(fps: f64, sample_rate: u32, position: isize) -> isize {
     calculate_samples_to_position(fps, sample_rate, position + 1)
         - calculate_samples_to_position(fps, sample_rate, position)
 }
 #[inline]
-pub fn calculate_samples_to_position(fps: f32, sample_rate: u32, position: isize) -> isize {
+pub fn calculate_samples_to_position(fps: f64, sample_rate: u32, position: isize) -> isize {
     if fps != 0.0 {
-        (position as f64 * sample_rate as f64 / fps as f64
-            + (if position < 0 { -0.5 } else { 0.5 })) as isize
+        (position as f64 * sample_rate as f64 / fps + (if position < 0 { -0.5 } else { 0.5 }))
+            as isize
     } else {
         0
     }
@@ -22,7 +22,7 @@ mod tests {
     #[rstest]
     #[case(0.0, 0, 0, 0)]
     fn calculate_frame_samples_works(
-        #[case] fps: f32,
+        #[case] fps: f64,
         #[case] sample_rate: u32,
         #[case] position: isize,
         #[case] expected: isize,
@@ -42,7 +42,7 @@ mod tests {
     #[case(30.0, 786432, 121, 3171942)]
     #[case(30.0, 786432, -121, -3171942)]
     fn calculate_samples_to_position_works(
-        #[case] fps: f32,
+        #[case] fps: f64,
         #[case] sample_rate: u32,
         #[case] position: isize,
         #[case] expected: isize,
